test/smoke: split out sousFlags
